requirements.txt
aiohttp==3.8.1
aiohttp-retry==2.4.6
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.7.5
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
asyncio==3.4.3
asyncpg==0.25.0
attrs==22.2.0
black==22.3.0
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.89.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
loguru==0.5.3
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.6.1
multidict==6.0.4
mypy-extensions==1.0.0
orjson==3.8.9
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pydantic==1.10.7
pyflakes==2.4.0
pytest-mock==3.10.0
python-dotenv==0.19.2
requests==2.27.1
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
tomli==2.0.1
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.17.0.post1
yarl==1.8.2
